Phase 2 commit 2e: hoist ComputerSystem, FileSystem, NetworkParams native-free code
FreeBsdComputerSystem abstract:
- Owns the manufModelSerialUuidVers supplier, dmidecode/lshal parsing,
  and the public getters.
- createFirmware() returns the now-hoisted FreeBsdFirmware.
- createBaseboard() returns UnixBaseboard (also moved to oshi-common).
- queryHostUuid() is the only abstract hook (kern.hostuuid fallback).

FreeBsdFileSystem abstract:
- Owns the full getFileStores() pipeline (geom part list, df -i, mount
  -p) and the path/volume include/exclude config keys.
- getOpenFileDescriptors/getMaxFileDescriptors delegate to two abstract
  hooks (kern.openfiles, kern.maxfiles).

FreeBsdOSFileStore moves to oshi-common alongside the abstract. Its
constructor now captures the owning FreeBsdFileSystem so
updateAttributes() can re-query against whichever JNA/FFM concrete
produced it, instead of the previous `new FreeBsdFileSystemJNA()`
back-reference.

FreeBsdNetworkParams abstract:
- getIpv4DefaultGateway/getIpv6DefaultGateway are pure `route get
  default` command-line.
- getDomainName/getHostName delegate to queryDomainName()/queryHostName()
  abstract hooks. queryHostName() returns null on native failure so the
  abstract falls back to AbstractNetworkParams.getHostName().

UnixBaseboard hoisted from oshi-core (oshi.hardware.platform.unix) to
oshi-common (oshi.hardware.common.platform.unix). It is pure-Java and
already shared between FreeBSD, NetBSD, OpenBSD, and Solaris; imports
updated in those three other ComputerSystem classes.
