+++ oshi-core/src/main/java/oshi/hardware/platform/unix/freebsd/FreeBsdSensorsJNA.java
@@ -15,9 +15,9 @@ import oshi.jna.ByRef.CloseableSizeTByReference;
 import oshi.jna.platform.unix.FreeBsdLibc;
 
 /**
- * JNA-backed FreeBSD sensors. The only native read is the per-CPU temperature from the {@code coretemp} kld module
- * via {@code sysctlbyname("dev.cpu.%d.temperature")}. Fan speeds and CPU voltage have no FreeBSD-supported source
- * and return the empty/zero defaults inherited from {@link FreeBsdSensors}.
+ * JNA-backed FreeBSD sensors. The only native read is the per-CPU temperature from the {@code coretemp} kld module via
+ * {@code sysctlbyname("dev.cpu.%d.temperature")}. Fan speeds and CPU voltage have no FreeBSD-supported source and
+ * return the empty/zero defaults inherited from {@link FreeBsdSensors}.
  */
 @ThreadSafe
 public class FreeBsdSensorsJNA extends FreeBsdSensors {
