+++ oshi-core/src/main/java/oshi/hardware/platform/unix/openbsd/OpenBsdComputerSystem.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 The OSHI Project Contributors
+ * Copyright 2021-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 package oshi.hardware.platform.unix.openbsd;
@@ -12,7 +12,7 @@ import oshi.annotation.concurrent.Immutable;
 import oshi.hardware.Baseboard;
 import oshi.hardware.Firmware;
 import oshi.hardware.common.AbstractComputerSystem;
-import oshi.hardware.platform.unix.UnixBaseboard;
+import oshi.hardware.common.platform.unix.UnixBaseboard;
 import oshi.util.Constants;
 import oshi.util.platform.unix.openbsd.OpenBsdSysctlUtil;
